lookup_kas: Fix some misc bugs in lookup_kas.c
The kas lookup methods had a few minor problems.  A regex error was causing some
symbols to get skipped in the parsing of /proc/kallsyms, leading to bogus
locations, and there was a minor memory leak of the name during lookup of new
symbols.  This patch fixes both of those problems.

Signed-off-by: Kenichi Maehashi <grep@maehashi.net>
Signed-off-by: Neil Horman <nhorman@tuxdriver.com>