Update distro to Xenial in the hopes that ubuntu will fix libpcap
libpcap in ubuntu is missing its .pc file, so autoconf fails.  They are
fixing it in newer releases but won't fix it in trusty, so augment the
configure script to use AC_CHECK_LIB if PKG_CHECK_MODULE fails.  Also
update to xenial in the travis-ci script so that we have a hope of
removing this in the future

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>