Update configure to fail if readline isn't installed
We had libreadline pkg check in configure hacked a bit.  Because
libreadline didn't have a .pc file for years, we fell back on just
adding the library as a linker option and letting the compile fail when
readline.h couldn't be found.  Now that readline 8.0 has been out for 2
years (which included the .pc file), I think we can safely test for it,
and fail if its not found

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>