Fix build without libtool
When libtool is not installed, automake calls the linker with the
library flags (`-llib`) _before_ the object files. In order to
successfully resolve symbols, the library flags need to come _after_ any
object file needing them. libtool does that correctly by itself, but
when automake cannot rely on libtool, we need to tell it about libraries
via `mybin_LDADD` instead of the global `AM_LDFLAGS`.